Move menus into it's own file
